+++ MyWork/Haritha/E8 - Interfaces and Classes/Result.cs
@@ -7,7 +7,7 @@ namespace E8___Interfaces_and_Classes
    public class Result
     {
         public bool Sumflag;
-        public int Sum;
+        public Double Sum;
         public Double Average, Median, Variance;
